Switch to webpack CLI for building libs; remove Gulp
