+++ packages/resourceful-action-creators/.babelrc
@@ -1,0 +1,11 @@
+{
+  "presets": ["latest", "stage-3"],
+  "plugins": [
+    "add-module-exports"
+  ],
+  "env": {
+    "test": {
+      "plugins": ["istanbul"]
+    }
+  }
+}
