+++ packages/resourceful-prop-types/package.json
@@ -2,7 +2,10 @@
   "name": "resourceful-prop-types",
   "version": "0.0.11",
   "description": "prop-types for Resourceful Redux",
-  "main": "dist/index.js",
+  "main": "dist/resourceful-prop-types.js",
+  "scripts": {
+    "build": "webpack"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/jmeas/resourceful-redux.git"
@@ -15,5 +18,14 @@
   "homepage": "https://resourceful-redux.js.org",
   "dependencies": {
     "prop-types": "^15.5.10"
+  },
+  "devDependencies": {
+    "babel-core": "6.14.0",
+    "babel-loader": "6.2.5",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-polyfill": "6.13.0",
+    "babel-preset-latest": "6.14.0",
+    "babel-preset-stage-3": "^6.17.0",
+    "webpack": "2.6.1"
   }
 }
