+++ packages/resourceful-redux/package.json
@@ -3,6 +3,9 @@
   "version": "0.0.11",
   "description": "Resource management for Redux.",
   "main": "dist/resourceful-redux.js",
+  "scripts": {
+    "build": "webpack"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/jmeas/resourceful-redux.git"
@@ -49,21 +52,6 @@
     "babel-polyfill": "6.13.0",
     "babel-preset-latest": "6.14.0",
     "babel-preset-stage-3": "^6.17.0",
-    "babel-register": "6.14.0",
-    "chai": "3.5.0",
-    "del": "2.2.2",
-    "gitbook-cli": "^2.3.0",
-    "glob": "7.0.6",
-    "gulp": "3.9.1",
-    "gulp-eslint": "3.0.1",
-    "json-loader": "0.5.4",
-    "lerna": "^2.0.0-rc.5",
-    "lodash": "^4.17.4",
-    "mocha": "3.0.2",
-    "pleaserc": "^2.3.1",
-    "sinon": "1.17.5",
-    "sinon-chai": "2.8.0",
-    "webpack": "2.6.1",
-    "webpack-stream": "^3.2.0"
+    "webpack": "2.6.1"
   }
 }
